litellm-proxy/pyproject.toml
[project]
name = "claude-cli-litellm"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["litellm[proxy]>=1.40", "python-dotenv>=1.0"]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "httpx>=0.27"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: requires running litellm proxy"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
